Refactor: 검색창을 컴포넌트화(playlist만 연결)
