+++ src/app/projectReview/page.tsx
@@ -1,10 +1,10 @@
 import ProjectReviewList from "../components/projectReview/projectReviewList";
-import ProjectReviewComponents from "../components/projectReview/reviewSearchForm";
+import ProjectSearchForm from "../components/searchList/SearchForm";
 
 function page() {
   return (
     <div>
-      <ProjectReviewComponents />
+      <ProjectSearchForm titleText="(아이콘) 참여한 프로젝트의 리뷰를 남겨보세요!" />
       <ProjectReviewList />
     </div>
   );
